stantard type UINT

+++ simulatorCore/src/scheduler.cpp
@@ -5,6 +5,7 @@
  *      Author: dom
  */
 
+#include <stdint.h>
 #include <iostream>
 #include <stdlib.h>
 #include "assert.h"
@@ -33,7 +34,7 @@ Scheduler::Scheduler() {
 	}
 
 	currentDate = 0;
-	maximumDate = UINT64_MAX; // no time limitation by default
+	maximumDate = UINT_MAX; // no time limitation by default
 
 	eventsMapSize = 0;
 	largestEventsMapSize = 0;
